WIP: Rename select contact activity
- Add application module
