+++ app/src/main/AndroidManifest.xml
@@ -8,11 +8,11 @@
             android:allowBackup="true"
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
-            android:name=".CursoAndroidApp"
+            android:name=".application.CursoAndroidApp"
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme.NoActionBar">
-        <activity android:name=".ListaContatosActivity">
+        <activity android:name=".application.ui.SelectContactActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
