+++ app/src/main/java/io/github/wellingtoncosta/cursoandroidcongressodeti/application/ui/SelectContactActivity.kt
@@ -1,9 +1,10 @@
-package io.github.wellingtoncosta.cursoandroidcongressodeti
+package io.github.wellingtoncosta.cursoandroidcongressodeti.application.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import io.github.wellingtoncosta.cursoandroidcongressodeti.R
 
-class ListaContatosActivity : AppCompatActivity() {
+class SelectContactActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
