+++ app/src/main/res/layout/activity_lista_contatos.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ListaContatosActivity">
+        tools:context=".application.ui.SelectContactActivity">
 
     <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/appbar"
